Split out gz_entity_filter implementation into a cpp file
